backend_fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
tensorflow>=2.16.1
scikit-learn==1.3.2
numpy>=1.26.0
pillow>=10.1.0
imagehash==4.3.1
pandas>=2.1.3
prophet==1.1.5
pulp==2.7.0
geopy==2.4.1
scipy>=1.11.4
nltk==3.8.1
requests==2.31.0
alembic==1.12.1

